+++ crates/scheduler/src/lib.rs
@@ -241,11 +241,11 @@ impl<B: Board> Scheduler<B> {
     #[cfg(any(feature = "pulley", feature = "wasm"))]
     pub fn run() -> ! {
         let mut scheduler = Self::new();
-        if let Err(error) = scheduler.start_applet() {
-            log::warn!("Failed to start applet: {}", error);
-        }
+        scheduler.start_applet();
         loop {
+            log::trace!("Flushing events.");
             scheduler.flush_events();
+            log::trace!("Processing applet.");
             scheduler.process_applet();
         }
     }
@@ -339,8 +339,16 @@ impl<B: Board> Scheduler<B> {
         }
     }
 
+    #[cfg(any(feature = "pulley", feature = "wasm"))]
+    fn start_applet(&mut self) {
+        match self.start_applet_() {
+            Ok(()) => (),
+            Err(e) => log::warn!("Failed to start applet: {}", e),
+        }
+    }
+
     #[cfg(feature = "pulley")]
-    fn start_applet(&mut self) -> Result<(), Error> {
+    fn start_applet_(&mut self) -> Result<(), Error> {
         // SAFETY: We stop the applet before installing a new one.
         let pulley = unsafe { <board::Applet<B> as board::applet::Api>::get()? };
         if pulley.is_empty() {
@@ -382,7 +390,7 @@ impl<B: Board> Scheduler<B> {
     }
 
     #[cfg(feature = "wasm")]
-    fn start_applet(&mut self) -> Result<(), Error> {
+    fn start_applet_(&mut self) -> Result<(), Error> {
         const MEMORY_SIZE: usize = memory_size();
         #[repr(align(16))]
         struct Memory([u8; MEMORY_SIZE]);
@@ -454,19 +462,13 @@ impl<B: Board> Scheduler<B> {
 
     /// Returns whether execution should resume.
     fn process_event(&mut self) -> bool {
-        let event = loop {
-            let applet = match self.applet.get() {
-                Some(x) => x,
-                None => return false,
-            };
-            match applet.pop() {
-                EventAction::Handle(event) => break event,
-                EventAction::Wait => self.wait_event(),
-                #[cfg(any(feature = "pulley", feature = "wasm"))]
-                EventAction::Reply => return true,
-            }
-        };
-        event::process(self, event);
+        let Some(applet) = self.applet.get() else { return false };
+        match applet.pop() {
+            EventAction::Handle(event) => event::process(self, event),
+            EventAction::Wait => self.wait_event(),
+            #[cfg(any(feature = "pulley", feature = "wasm"))]
+            EventAction::Reply => return true,
+        }
         false
     }